Remove dead code:
* Remove test runner v1 stuff
* Remove mutator plugin stuff
* Remove test framework plugin stuff
* Remove `SandboxPool` (was reimagined as `TestRunnerPool`)
* Remove `MutantTestMatcher` (was reimagined as `findMutantTestCoverage`)
