+++ packages/core/src/di/coreTokens.ts
@@ -2,23 +2,19 @@ export const cliOptions = 'cliOptions';
 export const configReader = 'configReader';
 export const configOptionsApplier = 'configOptionsApplier'; // TODO: Remove if unused post mutation switching
 export const inputFiles = 'inputFiles';
-export const initialRunResult = 'initialRunResult'; // TODO: Remove if unused post mutation switching
 export const dryRunResult = 'dryRunResult';
 export const transpiledFiles = 'transpiledFiles'; // TODO: Remove if unused post mutation switching
 export const files = 'files';
 export const mutants = 'mutants';
 export const mutantsWithTestCoverage = 'mutantsWithTestCoverage';
 export const mutantTranspileScheduler = 'mutantTranspileScheduler'; // TODO: Remove if unused post mutation switching
-export const sandboxPool = 'sandboxPool'; // TODO: Remove if unused post mutation switching
 export const temporaryDirectory = 'temporaryDirectory';
-export const testFramework = 'testFramework'; // TODO: Remove if unused post mutation switching
 export const timer = 'timer';
-export const timeOverheadMS = 'timeOverheadMS'; // TODO: Remove if unused post mutation switching
+export const timeOverheadMS = 'timeOverheadMS';
 export const loggingContext = 'loggingContext';
-export const mutationTestReportCalculator = 'mutationTestReportCalculator';
+export const mutationTestReportHelper = 'mutationTestReportHelper';
 export const transpiler = 'transpiler';
 export const sandbox = 'sandbox';
-export const sandboxIndex = 'sandboxIndex'; // TODO: Remove if unused post mutation switching
 export const sandboxWorkingDirectory = 'sandboxWorkingDirectory'; // TODO: Remove if unused post mutation switching
 export const testRunnerFactory = 'testRunnerFactory';
 export const testRunnerPool = 'testRunnerPool';
