+++ packages/core/src/di/factoryMethods.ts
@@ -1,12 +1,11 @@
 import { MutatorDescriptor, StrykerOptions } from '@stryker-mutator/api/core';
 import { Logger, LoggerFactoryMethod } from '@stryker-mutator/api/logging';
-import { commonTokens, Injector, OptionsContext, PluginKind, PluginResolver, tokens } from '@stryker-mutator/api/plugin';
+import { commonTokens, Injector, PluginResolver, tokens } from '@stryker-mutator/api/plugin';
 import { deepFreeze } from '@stryker-mutator/util';
 
 import { OptionsEditorApplier } from '../config';
-import TestFrameworkOrchestrator from '../TestFrameworkOrchestrator';
 
-import { coreTokens, PluginCreator, PluginLoader } from '.';
+import { coreTokens, PluginLoader } from '.';
 
 export function pluginResolverFactory(
   injector: Injector<{ [commonTokens.logger]: Logger; [coreTokens.pluginDescriptors]: readonly string[] }>
@@ -16,13 +15,6 @@ export function pluginResolverFactory(
   return pluginLoader;
 }
 pluginResolverFactory.inject = tokens(commonTokens.injector);
-
-export function testFrameworkFactory(
-  injector: Injector<OptionsContext & { [coreTokens.pluginCreatorTestFramework]: PluginCreator<PluginKind.TestFramework> }>
-) {
-  return injector.injectClass(TestFrameworkOrchestrator).determineTestFramework();
-}
-testFrameworkFactory.inject = tokens(commonTokens.injector);
 
 // eslint-disable-next-line @typescript-eslint/ban-types
 export function loggerFactory(getLogger: LoggerFactoryMethod, target: Function | undefined) {
