+++ packages/core/src/mutants/index.ts
@@ -1,0 +1,1 @@
+export * from './findMutantTestCoverage';
