+++ packages/core/test/unit/test-runner/RetryDecorator.spec.ts
@@ -9,8 +9,8 @@ import { expectErrored } from '@stryker-mutator/test-helpers/src/assertions';
 
 import ChildProcessCrashedError from '../../../src/child-proxy/ChildProcessCrashedError';
 import OutOfMemoryError from '../../../src/child-proxy/OutOfMemoryError';
-import RetryDecorator from '../../../src/test-runner-2/RetryDecorator';
-import TestRunnerDecorator from '../../../src/test-runner-2/TestRunnerDecorator';
+import RetryDecorator from '../../../src/test-runner/RetryDecorator';
+import TestRunnerDecorator from '../../../src/test-runner/TestRunnerDecorator';
 import currentLogMock from '../../helpers/logMock';
 
 describe(RetryDecorator.name, () => {
