+++ packages/core/test/unit/test-runner/TestRunnerDecorator.spec.ts
@@ -3,7 +3,7 @@ import { expect } from 'chai';
 import { TestRunner2 } from '@stryker-mutator/api/test_runner2';
 import { factory } from '@stryker-mutator/test-helpers';
 
-import TestRunnerDecorator from '../../../src/test-runner-2/TestRunnerDecorator';
+import TestRunnerDecorator from '../../../src/test-runner/TestRunnerDecorator';
 
 describe('TestRunnerDecorator', () => {
   let sut: TestRunnerDecorator;
